+++ products/contracts/conformance-contracts/src/checks/contracts.ts
@@ -6,9 +6,15 @@ import * as checks from "./checks";
 export type {
 	ConformanceCheckResult,
 	ConformanceCheckResultBase,
+	ParsedVersionedConformanceCheckResult,
+	VersionedConformanceCheckResult,
 } from "./checks";
 
 export const checksContracts = Object.freeze({
 	conformanceCheckResultSchema: checks.conformanceCheckResultSchema,
 	parseConformanceCheckResult: checks.parseConformanceCheckResult,
+	versionedConformanceCheckResultSchema:
+		checks.versionedConformanceCheckResultSchema,
+	parseVersionedConformanceCheckResult:
+		checks.parseVersionedConformanceCheckResult,
 });